FIXED STUPID F$#^*($ING BOOK CRASHING THE GAME
I wrestled with this damn thing for an hour and a bit since it started becoming a problem only later and I was messing with qud kissing too at the time, and pulled out alllll the damn wiring trying to find what was making the game crash on startup. turns out i forgot a backslash in the new quotebook, even though it worked the first time. unclear what happened. send me to hell. its fixed now at least